post_accreditation/requirements.txt
Flask-WTF==1.2.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
gunicorn==21.2.0
psycopg2-binary==2.9.9
Werkzeug==3.0.1
PyJWT==2.8.0
email-validator==2.1.0
APScheduler==3.10.4
python-dateutil==2.8.2
Flask-Mail==0.9.1
pyotp==2.9.0
qrcode[pil]==7.4.2
reportlab==4.0.8
Pillow==10.1.0
python-dotenv==1.0.0
Flask-Migrate==4.0.5
celery==5.3.4
redis==5.0.1
